refactor: decoupled demos from advanced_heaps
